==> mimic-notebook-analysis/0 lab - latest script.ipynb <==
data = "/Volumes/GoogleDrive/My Drive/TAU/Code/DrugLab/data"

import pandas as pd
import os

admits = pd.read_csv(os.path.join(data, "mimiciii/1.4/raw", "ADMISSIONS.csv.gz"))

## MIMIC Extract

mimic_extract_map = pd.read_csv(os.path.join("/Users/yanpavan/Desktop/", "itemid_to_variable_map.csv"))

mimic_extract_map_small = mimic_extract_map[["LEVEL2", "MIMIC LABEL", "ITEMID", "STATUS", "LINKSTO"]]
mimic_extract_map_small_lab = mimic_extract_map_small[mimic_extract_map_small["STATUS"]=="ready"]

d_m_l_doc = pd.read_csv(os.path.join(data, "mimiciii", "1.4","preprocessed", "mapping_med_itemid_doc.csv")).drop(columns=["Unnamed: 0"])

d_m_l_doc

lab_itemids = t_lab_itemids = {'ALT': {'Alanine aminotransferase': [50861, 769, 220644]},
 'ANA': {'Fraction inspired oxygen': [189]},
 'AST': {'Asparate aminotransferase': [50878, 770, 220587]},
 'Hemoglobin': {'Hemoglobin percent': [50852],
  'Hemoglobin C': [51224],
  'Hemoglobin F': [51225],
  'Hemoglobin A2': [51223],
  'Hemoglobin': [814, 220228, 51222, 50811]},
 'INR': {'Prothrombin time INR': [51237, 815, 1530, 227467]},
 'bilirubin': {'Bilirubin': [51465,
   50883,
   803,
   225651,
   50885,
   1538,
   848,
   225690,
   50884],
  'Bilirubin, Total, Pleural': [51049],
  'Bilirubin, Total, Body Fluid': [51028],
  'Bilirubin, Total, Ascites': [50838]},
 'calcium': {'Calcium': [786, 1522, 3746, 51029, 50893, 225625],
  'Calcium ionized': [50808, 816, 225667, 3766],
  'Calcium urine': [51066, 51077]},
 'creatinine': {'Creatinine': [791, 1525, 220615, 50912],
  'Creatinine ascites': [50841],
  'Creatinine body fluid': [51032],
  'Creatinine pleural': [51052],
  'Creatinine urine': [51082]},
 'glucose': {'Glucose': [50931,
   807,
   811,
   1529,
   50809,
   3745,
   225664,
   220621,
   226537],
  'Glucose urine': [51478],
  'Glucose, CSF': [51014],
  'Estimated Actual Glucose': [51529],
  'Glucose, Urine': [51084],
  'Glucose, Pleural': [51053],
  'Glucose, Joint Fluid': [51022],
  'Glucose, Ascites': [50842],
  'Glucose, Body Fluid': [51034]},
 'lactic acid': {'Lactic acid': [818, 225668, 1531]},
 'magnesium': {'Magnesium': [50960], 'Magnesium, Urine': [51088]},
 'platelets': {'Platelets': [51265, 828, 227457], 'Large Platelets': [51240]},
 'potassium': {'Potassium': [829, 1535, 227464, 50971, 50822],
  'Potassium serum': [227442],
  'Potassium, Body Fluid': [51041],
  'Potassium, Pleural': [51057],
  'Potassium, Stool': [51064],
  'Potassium, Urine': [51097],
  'Potassium, Ascites': [50847]},
 'sodium': {'Sodium': [837, 1536, 220645, 226534, 50983, 50824],
  'Sodium, Ascites': [50848],
  'Sodium, Body Fluid': [51042],
  'Sodium, Pleural': [51058],
  'Sodium, Stool': [51065],
  'Sodium, Urine': [51100]},
 'Uric acid': {'Uric Acid': [51007], 'Uric Acid, Urine': [51105]},
 'B12': {'Vitamin B12': [51010]},
 'prolactin': {'Prolactin': [50973]},
 'Amylase': {'Amylase': [50867], 'Amylase, Ascites': [50836], 'Amylase, Body Fluid': [51026],'Amylase, Joint Fluid': [51020],'Amylase, Pleural': [51047], 'Amylase, Urine': [51072]},
 'Lipase': {'Lipase': [50956], 'Lipase, Ascites': [50844], 'Lipase, Body Fluid': [51036]},
 'Aptt': {'PTT': [825, 1533, 227466, 51275]}}

lab_itemids['Hematocrit']  = {
    'Hematocrit' : [813, 220545, 51221, 50810]
}
lab_itemids['Red blood cell'] = {
    'Red blood cell': [51279, 833]
}
lab_itemids['Albumin'] = {
    'Albumin': [50862, 772, 1521, 227456]
}
lab_itemids['Magnesium'] = {
    'Magnesium': [821, 1532, 220635, 50960]
}

### Manual Mapping

mimic_extract_map_small_lab[mimic_extract_map_small_lab["LEVEL2"]=="Red blood cell count"][["LEVEL2", "ITEMID"]].set_index("LEVEL2").ITEMID.to_list()

mimic_extract_map_small_lab[mimic_extract_map_small_lab["LEVEL2"]=="Hemoglobin"][["LEVEL2", "ITEMID"]].set_index("LEVEL2").ITEMID.to_list()

mimic_extract_map_small_lab[mimic_extract_map_small_lab["LEVEL2"]=="Bilirubin"][["LEVEL2", "ITEMID"]].set_index("LEVEL2").ITEMID.to_list()

mimic_extract_map_small_lab[mimic_extract_map_small_lab["LEVEL2"]=="Creatinine urine"][["LEVEL2", "ITEMID"]].set_index("LEVEL2").ITEMID.to_list()

mimic_extract_map_small_lab[mimic_extract_map_small_lab["LEVEL2"]=="Calcium"][["LEVEL2", "ITEMID"]].set_index("LEVEL2").ITEMID.to_list()

mimic_extract_map_small_lab[mimic_extract_map_small_lab["LEVEL2"]=="Glucose"][["LEVEL2", "ITEMID"]].set_index("LEVEL2").ITEMID.to_list()

mimic_extract_map_small_lab[mimic_extract_map_small_lab["LEVEL2"]=="Platelets"][["LEVEL2", "ITEMID"]].set_index("LEVEL2").ITEMID.to_list()

mimic_extract_map_small_lab[mimic_extract_map_small_lab["LEVEL2"]=="Potassium"][["LEVEL2", "ITEMID"]].set_index("LEVEL2").ITEMID.to_list()

mimic_extract_map_small_lab[mimic_extract_map_small_lab["LEVEL2"]=="Sodium"][["LEVEL2", "ITEMID"]].set_index("LEVEL2").ITEMID.to_list()

mimic_extract_map_small_lab[mimic_extract_map_small_lab["LEVEL2"]=="Partial thromboplastin time"]

mimic_extract_map_small_lab[mimic_extract_map_small_lab["LEVEL2"]=="Partial thromboplastin time"][["LEVEL2", "ITEMID"]].set_index("LEVEL2").ITEMID.to_list()

mimic_extract_map_small_lab[mimic_extract_map_small_lab["LEVEL2"]=="Hematocrit"]

mimic_extract_map_small_lab[mimic_extract_map_small_lab["LEVEL2"]=="Hematocrit"][["LEVEL2", "ITEMID"]].set_index("LEVEL2").ITEMID.to_list()

mimic_extract_map_small_lab[mimic_extract_map_small_lab["LEVEL2"]=="Albumin"][["LEVEL2", "ITEMID"]].set_index("LEVEL2").ITEMID.to_list()

mimic_extract_map_small_lab[mimic_extract_map_small_lab["LEVEL2"]=="Magnesium"][["LEVEL2", "ITEMID"]].set_index("LEVEL2").ITEMID.to_list()

### Final Mapping

final_mapping_lab_itemids = {v2:k1 for k, v in lab_itemids.items() for k1, v1 in v.items() for v2 in v1}

final_itemids_list = list(final_mapping_lab_itemids.keys())


## Labevents

labevents = pd.read_csv(os.path.join(data, "mimiciii/1.4/preprocessed", "lab_patient_data.csv"))

labevents = labevents[labevents.ITEMID.isin(final_itemids_list)]
labevents["MIMICExtractName"] = labevents.apply(lambda r: final_mapping_lab_itemids[r["ITEMID"]], axis=1)

labevents["TABLE"] = labevents.apply(lambda r: "LABEVENTS", axis=1)

patients = pd.read_csv(os.path.join(data, "mimiciii/1.4", "raw/PATIENTS.csv.gz"))
labevents = labevents.rename(columns={"SUBJECT_ID_x":"SUBJECT_ID"})
labevents = pd.merge(labevents, patients, how="inner", on="SUBJECT_ID")

labevents["DOB"] = pd.to_datetime(labevents["DOB"])
labevents["ADMITTIME"] = pd.to_datetime(labevents["ADMITTIME"])

labevents['AGE'] = labevents.apply(lambda r: round((r['ADMITTIME'].to_pydatetime()-r['DOB'].to_pydatetime()).days/365, 0), axis=1)

labevents[['ADMISSION_LOCATION',
 'ADMISSION_TYPE',
 'ADMITTIME',
 'AGE',
 'CHARTTIME',
 'DEATHTIME',
 'DIAGNOSIS',
 'DISCHARGE_LOCATION',
 'DISCHTIME',
 'DOB',
 'DOD',
 'DOD_HOSP',
 'DOD_SSN',
 'EDOUTTIME',
 'EDREGTIME',
 'ETHNICITY',
 'EXPIRE_FLAG',
 'GENDER',
 'HADM_ID',
 'HAS_CHARTEVENTS_DATA',
 'HOSPITAL_EXPIRE_FLAG',
 'INSURANCE',
 'ITEMID',
 'LANGUAGE',
 'MARITAL_STATUS',
 'MIMICExtractName',
 'RELIGION',
 'ROW_ID_x',
 'SUBJECT_ID',
 'TABLE',
 'VALUE',
 'VALUENUM',
 'VALUEUOM']].to_csv(os.path.join(data, "mimiciii/1.4/preprocessed", "lab_patient_data_with_mimic_extract_names.csv"))

labevents = pd.read_csv(os.path.join(data, "mimiciii/1.4/preprocessed", "lab_patient_data_with_mimic_extract_names.csv"))

## Chartevents

res_paths = []
count = 0
with pd.read_csv(os.path.join(data, "mimiciii/1.4/raw", "CHARTEVENTS.csv.gz"), chunksize=50_000_000) as reader:
    for chunk in reader:
        print(count, chunk.shape)
        path = os.path.join(data, "mimiciii", "1.4","preprocessed", "CHARTEVENTS", f"chartevents_with_mimic_extract_{count}.csv.gz")
        chunk = chunk[chunk.ITEMID.isin(final_itemids_list)]
        t = chunk.apply(lambda r: final_mapping_lab_itemids[r["ITEMID"]], axis=1)
        if t.shape[0]>0:
            chunk["MIMICExtractName"] = t
            chunk.to_csv(path)
            print(count, chunk.shape)
            res_paths.append(path)
        else:
            print(count, 0)    
        count += 1


result = pd.concat([pd.read_csv(f) for f in res_paths], ignore_index=True)

result["TABLE"] = result.apply(lambda r: "CHARTEVENTS", axis=1)

t_r = pd.merge(result, admits, how="inner", on="HADM_ID")

result = t_r.copy()

result = result.rename(columns={"SUBJECT_ID_x":"SUBJECT_ID"})
result = pd.merge(result, patients, how="inner", on="SUBJECT_ID")

result["DOB"] = pd.to_datetime(result["DOB"])
result["ADMITTIME"] = pd.to_datetime(result["ADMITTIME"])

result['AGE'] = result.apply(lambda r: round((r['ADMITTIME'].to_pydatetime()-r['DOB'].to_pydatetime()).days/365, 0), axis=1)

result[['ADMISSION_LOCATION',
 'ADMISSION_TYPE',
 'ADMITTIME',
 'AGE',
 'CHARTTIME',
 'DEATHTIME',
 'DIAGNOSIS',
 'DISCHARGE_LOCATION',
 'DISCHTIME',
 'DOB',
 'DOD',
 'DOD_HOSP',
 'DOD_SSN',
 'EDOUTTIME',
 'EDREGTIME',
 'ETHNICITY',
 'EXPIRE_FLAG',
 'GENDER',
 'HADM_ID',
 'HAS_CHARTEVENTS_DATA',
 'HOSPITAL_EXPIRE_FLAG',
 'INSURANCE',
 'ITEMID',
 'LANGUAGE',
 'MARITAL_STATUS',
 'MIMICExtractName',
 'RELIGION',
 'ROW_ID_x',
 'SUBJECT_ID',
 'SUBJECT_ID_y',
 'TABLE',
 'VALUE',
 'VALUENUM',
 'VALUEUOM']].to_csv(os.path.join(data, "mimiciii", "1.4","preprocessed", "CHARTEVENTS", "chartevents_with_mimic_extract_lab.csv.gz"))

## Merge

t_r = result[['ADMISSION_LOCATION',
 'ADMISSION_TYPE',
 'ADMITTIME',
 'AGE',
 'CHARTTIME',
 'DEATHTIME',
 'DIAGNOSIS',
 'DISCHARGE_LOCATION',
 'DISCHTIME',
 'DOB',
 'DOD',
 'DOD_HOSP',
 'DOD_SSN',
 'EDOUTTIME',
 'EDREGTIME',
 'ETHNICITY',
 'EXPIRE_FLAG',
 'GENDER',
 'HADM_ID',
 'HAS_CHARTEVENTS_DATA',
 'HOSPITAL_EXPIRE_FLAG',
 'INSURANCE',
 'ITEMID',
 'LANGUAGE',
 'MARITAL_STATUS',
 'MIMICExtractName',
 'RELIGION',
 'SUBJECT_ID',
#  'TABLE',
 'VALUE',
 'VALUENUM',
 'VALUEUOM']]

t_l = labevents[['ADMISSION_LOCATION',
 'ADMISSION_TYPE',
 'ADMITTIME',
 'AGE',
 'CHARTTIME',
 'DEATHTIME',
 'DIAGNOSIS',
 'DISCHARGE_LOCATION',
 'DISCHTIME',
 'DOB',
 'DOD',
 'DOD_HOSP',
 'DOD_SSN',
 'EDOUTTIME',
 'EDREGTIME',
 'ETHNICITY',
 'EXPIRE_FLAG',
 'GENDER',
 'HADM_ID',
 'HAS_CHARTEVENTS_DATA',
 'HOSPITAL_EXPIRE_FLAG',
 'INSURANCE',
 'ITEMID',
 'LANGUAGE',
 'MARITAL_STATUS',
 'MIMICExtractName',
 'RELIGION',
 'SUBJECT_ID',
#  'TABLE',
 'VALUE',
 'VALUENUM',
 'VALUEUOM']]

t_c_l = pd.concat([t_l, t_r]).drop_duplicates(keep="first")

t_c_l = t_c_l[t_c_l["AGE"]<100]
t_c_l = t_c_l[t_c_l["AGE"]>0]

final_labs = t_c_l.groupby(["HADM_ID", "ADMITTIME", "CHARTTIME", "VALUENUM", "MIMICExtractName"]).nth(0).reset_index()

final_labs = final_labs.dropna(subset=['VALUENUM'])

final_labs.to_csv(os.path.join(data, "mimiciii/1.4/preprocessed", "lab_patient_data_mimic_extract_2.csv"))
