==> mimic-notebook-analysis/1 med.ipynb <==
data = "/Volumes/GoogleDrive/My Drive/TAU/Code/DrugLab/data"

import pandas as pd
import os

admits = pd.read_csv(os.path.join(data, "mimiciii/1.4/raw", "ADMISSIONS.csv.gz"))

admits['SUBJECT_ID'].unique().shape

inputevents_mv = pd.read_csv(os.path.join(data, "mimiciii/1.4/raw", "INPUTEVENTS_MV.csv.gz"))

inputevents_mv

inputevents_mv = pd.merge(inputevents_mv, admits, how="inner", on="HADM_ID")

inputevents_mv

def change_col_to_datetime(inputevents_mv, feature):
    inputevents_mv[feature] = pd.to_datetime(inputevents_mv[feature])
    return inputevents_mv
inputevents_mv = change_col_to_datetime(inputevents_mv, 'ADMITTIME')
inputevents_mv = change_col_to_datetime(inputevents_mv, 'ENDTIME')
inputevents_mv = change_col_to_datetime(inputevents_mv, 'STARTTIME')
inputevents_mv['MedTimeFromAdmit'] = inputevents_mv['ENDTIME']-inputevents_mv['ADMITTIME']

med_data = pd.read_csv(os.path.join(data, "mimiciii/1.4/raw", "D_ITEMS.csv.gz"))

med_data = med_data[med_data["DBSOURCE"]=="metavision"]

inputevents_mv = pd.merge(inputevents_mv, med_data, how="inner", on="ITEMID")

inputevents_mv

inputevents_mv.to_csv(os.path.join(data, "mimiciii/1.4/preprocessed", "med_patient_data.csv"))

inputevents_mv.describe()

def transform_column(col):
    return pd.to_datetime(col)
def gen_cols(med_pat):
    cols = ['STARTTIME', 'ENDTIME', 'ADMITTIME']
    for col in cols:
        med_pat[col] = transform_column(med_pat[col])
    med_pat.drop(columns=["MedTimeFromAdmit"])
    med_pat['MedTimeFromAdmit'] = med_pat['ENDTIME']-med_pat['ADMITTIME']
    return med_pat

med_pat = pd.read_csv(os.path.join(data, "mimiciii/1.4/preprocessed", "med_patient_data.csv"))

med_pat = gen_cols(med_pat)

med_pat.dtypes

med_pat.columns

med_pat.to_csv(os.path.join(data, "mimiciii/1.4/preprocessed", "med_patient_data.csv"))
