==> mimic-notebook-analysis/2 lab_med1_med2.ipynb <==
data = "/Volumes/GoogleDrive/My Drive/TAU/Code/DrugLab/data"

import pandas as pd
import os

## Lab Patient Data

l_p_df = pd.read_csv(os.path.join(data, "mimiciii/1.4/preprocessed", "lab_patient_data.csv"))

l_p_df = l_p_df.drop(columns=["Unnamed: 0"])

l_p_df["LabTimeFromAdmit"] = pd.to_datetime(l_p_df["CHARTTIME"])-pd.to_datetime(l_p_df["ADMITTIME"])

l_p_df

## Medication Patient Data

m_p_df = pd.read_csv(os.path.join(data, "mimiciii/1.4/preprocessed", "med_patient_data.csv"))

m_p_df = m_p_df.drop(columns=["Unnamed: 0"]).rename(columns={"SUBJECT_ID_x":"SUBJECT_ID"})

m_p_df["MedTimeFromAdmit"] = pd.to_datetime(m_p_df["ENDTIME"])-pd.to_datetime(m_p_df["ADMITTIME"])

m_p_df = m_p_df.sort_values(by="MedTimeFromAdmit")

m_p_df

## Filter Meds

### Med 1

med1 = m_p_df.groupby(["HADM_ID", "ITEMID"]).nth(0).reset_index()

med1 = med1.sort_values("MedTimeFromAdmit")

# med1[['HADM_ID', 'ITEMID', 'ROW_ID_x', 'SUBJECT_ID_x', 'ICUSTAY_ID',
#        'STARTTIME', 'ENDTIME', 'AMOUNT', 'AMOUNTUOM', 'RATE', 'RATEUOM', 'ADMITTIME', 'DISCHTIME', 'DEATHTIME', 'MedTimeFromAdmit']]

med1 = med1[med1["MedTimeFromAdmit"].dt.total_seconds()>0]

med1 = med1.sort_values(by=["ADMITTIME"]).groupby(["SUBJECT_ID", "ITEMID"]).nth(0).reset_index().sort_values(by=["MedTimeFromAdmit"])

med1

### Med 2

med2 = m_p_df.groupby(["HADM_ID", "ITEMID"]).nth(1).reset_index()

med2 = med2.sort_values("MedTimeFromAdmit")

med2 = med2[med2["MedTimeFromAdmit"].dt.total_seconds()>0]

med2 = med2.sort_values(by=["ADMITTIME"]).groupby(["SUBJECT_ID", "ITEMID"]).nth(0).reset_index().sort_values(by=["MedTimeFromAdmit"])

med2

## Filter Labs

labs = l_p_df[l_p_df["HADM_ID"].isin(med1["HADM_ID"])]

labs = labs[labs['LabTimeFromAdmit'].dt.total_seconds()>0]

labs = labs.sort_values("LabTimeFromAdmit")

labs = labs[labs['VALUENUM'].notna()]

labs

## Pairs

def check_med2(row):
    if row["HADM_ID"] in med2["HADM_ID"]:
        if row["ITEMID"] in med2[med2["HADM_ID"]==row["HADM_ID"]]["ITEMID"]:
            return True
    return False

def get_med2(row):
    temp = med2[med2["HADM_ID"]==row["HADM_ID"]] 
    return temp[temp["ITEMID"]==row["ITEMID"]].iloc[0]

def get_lab_data(row, type):
    lab_vals = labs[labs["HADM_ID"]==row["HADM_ID"]]
    if type=="mean_before":
        lab_vals = lab_vals[lab_vals["LabTimeFromAdmit"].dt.total_seconds()<row["MedTimeFromAdmit"].total_seconds()]
        if lab_vals.shape[0]==0:
            return -1
        return lab_vals.groupby("HADM_ID").mean()["VALUENUM"].iloc[0]
    elif type=="mean_after":
        med2_bool = check_med2(row)
        lab_vals = lab_vals[lab_vals["LabTimeFromAdmit"].dt.total_seconds()>row["MedTimeFromAdmit"].total_seconds()]
        if med2_bool:
            med2_val = get_med2(row)
            lab_vals = lab_vals[lab_vals["LabTimeFromAdmit"].dt.total_seconds()<med2_val["MedTimeFromAdmit"].total_seconds()]
            if lab_vals.shape[0]==0:
                return -1
        return lab_vals.groupby("HADM_ID").mean()["VALUENUM"].iloc[0]

t1 = med1.apply(lambda row : get_lab_data(row, "mean_before"), axis=1)

med1["LabBefore"] = t1

med1 = med1[med1["LabBefore"]!=-1]

med1 = med1.drop(columns=["Unnamed: 0.1", "Unnamed: 0.2"])

med1

med1.to_csv(os.path.join(data, "mimiciii/1.4/preprocessed", "lab_before.csv"))
